Classe FabMovementUtil.java
