+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name" translatable="false">FinToday-0.1.4.3344</string>
+    <string name="app_name" translatable="false">FinToday-0.1.4.3345</string>
     <string name="navigation_drawer_open">Abrir menu</string>
     <string name="navigation_drawer_close">Fechar menu</string>
     <string name="valor_desp">Valor Desp</string>
